requirements.txt
pandas>=2.0.0
pyyaml>=6.0
rich>=13.0.0
python-dateutil>=2.8.2
ipaddress>=1.0.23
python-evtx>=1.0.0
lxml>=4.9.0
python-registry>=1.3.1
pytz>=2023.3
xmltodict>=0.13.0
